backend/requirements.txt
# Web框架和相关依赖
fastapi>=0.110.0
uvicorn[standard]>=0.27.1
python-multipart>=0.0.9

# 数据验证和配置
pydantic>=2.6.3
pydantic-settings>=2.2.1
python-dotenv>=1.0.1

# 数据库相关
sqlalchemy>=2.0.27
alembic>=1.13.1
aiomysql>=0.2.0
pymysql>=1.1.0

# 缓存
redis>=4.2.0
aioredis>=2.0.1

# AI模型客户端
openai>=1.12.0
httpx>=0.27.0

# 日志
loguru>=0.7.2

# 工具包
python-jose[cryptography]>=3.3.0  # JWT相关
passlib[bcrypt]>=1.7.4  # 密码哈希
python-dateutil>=2.8.2  # 日期处理
email-validator>=2.1.0  # 邮件验证

# 开发工具
black>=24.2.0  # 代码格式化
isort>=5.13.2  # import排序
flake8>=7.0.0  # 代码检查
pytest>=8.0.2  # 测试框架
pytest-asyncio>=0.23.5  # 异步测试支持
pytest-cov>=4.1.0  # 测试覆盖率

# 文件处理
python-magic>=0.4.27
python-docx>=0.8.11
PyPDF2>=3.0.0
EbookLib>=0.18
markdown>=3.4
pillow>=10.0.0

# 文件存储
aiofiles>=23.1.0
python-multipart>=0.0.9

# OCR支持
pytesseract>=0.3.10